Fix: command status error float/int

+++ pomodoro-client.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import math
 from threading import Thread
 from subprocess import Popen
@@ -21,8 +23,8 @@ def get_pomodoro_proxy():
 
 def format_time(seconds):
     return "{minutes:02d}:{seconds:02d}".format(
-        minutes=math.floor(seconds / 60),
-        seconds=round(seconds % 60)
+        minutes=int(math.floor(seconds / 60)),
+        seconds=int(round(seconds % 60))
     )
